% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_format.R
\name{wide_format}
\alias{wide_format}
\title{Pivoting qualcoder data into one columne per code and category `(theme)`}
\usage{
wide_format(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
A data.frame with one column per code and category. Each row represent one participant, and each code and category are coded with 1 and 0.
}
\description{
Pivoting qualcoder data into one columne per code and category `(theme)`
}
